Redirects users form editing shops they don't own
